+++ source/extensions/lightspeed.trex.mesh_properties.shared.widget/lightspeed/trex/mesh_properties/shared/widget/tests/e2e/test_widget.py
@@ -1,3 +1,4 @@
+# noqa PLC0302
 """
 * SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 * SPDX-License-Identifier: Apache-2.0
@@ -15,6 +16,9 @@
 * limitations under the License.
 """
 
+import shutil
+import tempfile
+
 import carb.input
 import omni.ui as ui
 import omni.usd
@@ -24,7 +28,9 @@ from lightspeed.layer_manager.core import LayerManagerCore as _LayerManagerCore
 from lightspeed.layer_manager.core import LayerType as _LayerType
 from lightspeed.trex.mesh_properties.shared.widget import SetupUI as _MeshPropertiesWidget
 from lightspeed.trex.selection_tree.shared.widget import SetupUI as _SelectionTreeWidget
+from omni.flux.utils.common import path_utils as _path_utils
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
+from omni.flux.validator.factory import BASE_HASH_KEY
 from omni.kit import ui_test
 from omni.kit.test.async_unittest import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, open_stage, wait_stage_loading
@@ -97,6 +103,16 @@ class TestSelectionTreeWidget(AsyncTestCase):
         mesh_property_wid.destroy()
         selection_wid.destroy()
         window.destroy()
+
+        # destroy prompt dialogs to avoid unwanted references
+        for other_window in ui.Workspace.get_windows():
+            if other_window.title in {
+                _constants.ASSET_NEED_INGEST_WINDOW_TITLE,
+                _constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE,
+            }:
+                prompt_dialog_window = ui_test.find(other_window.title)
+                if prompt_dialog_window:
+                    prompt_dialog_window.widget.destroy()
 
     async def test_select_one_prim_mesh(self):
         # setup
@@ -362,7 +378,7 @@ class TestSelectionTreeWidget(AsyncTestCase):
         # It takes a while for the tree to update
         await ui_test.human_delay(10)
 
-        asset_path = _get_test_data("usd/ingested_assets/output/good/cube.usda")
+        asset_path = _get_test_data("usd/project_example/ingested_assets/output/good/cube.usda")
 
         # This doesn't work! Because there is a combobox stacked over the input field, on the CI, it will click on
         # the arrow of the combobox and input nothing!
@@ -383,6 +399,266 @@ class TestSelectionTreeWidget(AsyncTestCase):
         # the new asset give us 4 prims
         item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
         self.assertEquals(len(item_prims), 5)
+
+        await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_file_picker_outside_project_dir_cancel(self):
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda.meta"), temp_dir)
+
+            # Setup
+            _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+            # Select
+            usd_context = omni.usd.get_context()
+            usd_context.get_selection().set_selected_prim_paths(
+                ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+            )
+            await ui_test.human_delay(human_delay_speed=3)
+
+            item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+            self.assertEquals(len(item_prims), 3)
+            await item_prims[0].click()
+
+            # Test replace a mesh with a new one
+            replace_ref_open_folder = ui_test.find(
+                f"{_window.title}//Frame/**/Image[*].identifier=='replace_ref_open_folder'"
+            )
+            self.assertIsNotNone(replace_ref_open_folder)
+
+            await replace_ref_open_folder.click()
+            await ui_test.human_delay(1)
+
+            window_name = "Select a reference file"
+
+            # The file picker window should now be opened (0 < len(widgets))
+            self.assertLess(0, len(ui_test.find_all(f"{window_name}//Frame/**/*")))
+
+            select_button = ui_test.find(f"{window_name}//Frame/**/Button[*].text=='Select'")
+            toolbar_field = ui_test.find(
+                f"{window_name}//Frame/**/Rectangle[*].style_type_name_override=='ToolBar.Field'"
+            )
+
+            self.assertIsNotNone(select_button)
+            self.assertIsNotNone(toolbar_field)
+
+            # It takes a while for the tree to update
+            await ui_test.human_delay(10)
+
+            # Select file in filepicker
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            field_position = toolbar_field.position
+            await toolbar_field.click(pos=ui_test.Vec2(field_position.x + 1, field_position.y + 1))
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.TAB)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_char_press(asset_path)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.ENTER)
+            await select_button.click()
+            await ui_test.human_delay()
+
+        # Make sure that the ingestion window did not show
+        ignore_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNone(ignore_ingestion_button)
+        self.assertIsNone(cancel_ingestion_button)
+
+        # Make sure the "Copy Asset" and "Cancel" buttons exist
+        copy_external_asset_button = ui_test.find(
+            f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_external_asset_button = ui_test.find(
+            f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNotNone(copy_external_asset_button)
+        self.assertIsNotNone(cancel_external_asset_button)
+
+        # Cancel import and ensure the external asset was not imported
+        await cancel_external_asset_button.click()
+        await ui_test.human_delay(5)
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+
+        await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_file_picker_outside_project_dir_copy_and_re_ref(self):
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda.meta"), temp_dir)
+
+            # Setup
+            _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+            # Select
+            usd_context = omni.usd.get_context()
+            usd_context.get_selection().set_selected_prim_paths(
+                ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+            )
+            await ui_test.human_delay(human_delay_speed=3)
+
+            item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+            self.assertEquals(len(item_prims), 3)
+            await item_prims[0].click()
+
+            # Test replace a mesh with a new one
+            replace_ref_open_folder = ui_test.find(
+                f"{_window.title}//Frame/**/Image[*].identifier=='replace_ref_open_folder'"
+            )
+            self.assertIsNotNone(replace_ref_open_folder)
+
+            await replace_ref_open_folder.click()
+            await ui_test.human_delay(1)
+
+            window_name = "Select a reference file"
+
+            # The file picker window should now be opened (0 < len(widgets))
+            self.assertLess(0, len(ui_test.find_all(f"{window_name}//Frame/**/*")))
+
+            select_button = ui_test.find(f"{window_name}//Frame/**/Button[*].text=='Select'")
+            toolbar_field = ui_test.find(
+                f"{window_name}//Frame/**/Rectangle[*].style_type_name_override=='ToolBar.Field'"
+            )
+
+            self.assertIsNotNone(select_button)
+            self.assertIsNotNone(toolbar_field)
+
+            # It takes a while for the tree to update
+            await ui_test.human_delay(10)
+
+            # Select file in filepicker
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            field_position = toolbar_field.position
+            await toolbar_field.click(pos=ui_test.Vec2(field_position.x + 1, field_position.y + 1))
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.TAB)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_char_press(asset_path)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.ENTER)
+            await select_button.click()
+            await ui_test.human_delay()
+
+            # Make sure that the ingestion window did not show
+            ignore_ingestion_button = ui_test.find(
+                f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+            )
+            cancel_ingestion_button = ui_test.find(
+                f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+            )
+            self.assertIsNone(ignore_ingestion_button)
+            self.assertIsNone(cancel_ingestion_button)
+
+            # Make sure the "Copy Asset" and "Cancel" buttons exist
+            copy_external_asset_button = ui_test.find(
+                f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+            )
+            cancel_external_asset_button = ui_test.find(
+                f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+            )
+            self.assertIsNotNone(copy_external_asset_button)
+            self.assertIsNotNone(cancel_external_asset_button)
+
+            # Copy the asset
+            await copy_external_asset_button.click()
+            await ui_test.human_delay(50)
+
+            # Make sure that new ref exists
+            item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+            self.assertEqual(len(item_prims), 5)
+
+            # Check that the reference is from a new internal copy and not the original external asset
+            item_ref = item_prims[0].widget
+            self.assertEqual(item_ref.text, "cube.usda")
+            self.assertEqual(item_ref.tooltip, "./assets/ingested/cube.usda")
+
+            # Make sure the metadata matches
+            self.assertTrue(_path_utils.hash_match_metadata(file_path=asset_path, key=BASE_HASH_KEY))
+
+            # Delete the newly created project_example/assets/ingested subdirectory and its contents
+            shutil.rmtree(_get_test_data(f"usd/project_example/{_constants.REMIX_INGESTED_ASSETS_FOLDER}"))
+
+            await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_file_picker_outside_project_dir_no_metadata_cancel(self):
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+
+            # Setup
+            _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+            # Select
+            usd_context = omni.usd.get_context()
+            usd_context.get_selection().set_selected_prim_paths(
+                ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+            )
+            await ui_test.human_delay(human_delay_speed=3)
+
+            item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+            self.assertEquals(len(item_prims), 3)
+            await item_prims[0].click()
+
+            # Test replace a mesh with a new one
+            replace_ref_open_folder = ui_test.find(
+                f"{_window.title}//Frame/**/Image[*].identifier=='replace_ref_open_folder'"
+            )
+            self.assertIsNotNone(replace_ref_open_folder)
+
+            await replace_ref_open_folder.click()
+            await ui_test.human_delay(1)
+
+            window_name = "Select a reference file"
+
+            # The file picker window should now be opened (0 < len(widgets))
+            self.assertLess(0, len(ui_test.find_all(f"{window_name}//Frame/**/*")))
+
+            select_button = ui_test.find(f"{window_name}//Frame/**/Button[*].text=='Select'")
+            toolbar_field = ui_test.find(
+                f"{window_name}//Frame/**/Rectangle[*].style_type_name_override=='ToolBar.Field'"
+            )
+
+            self.assertIsNotNone(select_button)
+            self.assertIsNotNone(toolbar_field)
+
+            # It takes a while for the tree to update
+            await ui_test.human_delay(10)
+
+            # Select file in filepicker
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            field_position = toolbar_field.position
+            await toolbar_field.click(pos=ui_test.Vec2(field_position.x + 1, field_position.y + 1))
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.TAB)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_char_press(asset_path)
+            await ui_test.human_delay(1)
+            await ui_test.emulate_keyboard_press(KeyboardInput.ENTER)
+            await select_button.click()
+            await ui_test.human_delay()
+
+        # Make sure that the ingestion popup appeared - there should be no "ok" button
+        ignore_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNone(ignore_ingestion_button)
+        self.assertIsNotNone(cancel_ingestion_button)
+
+        # Cancel import and ensure the external asset was not imported
+        await cancel_ingestion_button.click()
+        await ui_test.human_delay(5)
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
 
         await self.__destroy(_window, _selection_wid, _mesh_property_wid)
 
@@ -469,7 +745,7 @@ class TestSelectionTreeWidget(AsyncTestCase):
         await mesh_ref_field.click()
         await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.A, carb.input.KEYBOARD_MODIFIER_FLAG_CONTROL)
         await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.DEL)
-        asset_path = _get_test_data("usd/ingested_assets/output/good/cube.usda")
+        asset_path = _get_test_data("usd/project_example/ingested_assets/output/good/cube.usda")
         await mesh_ref_field.input(asset_path, end_key=KeyboardInput.ENTER)
         await ui_test.human_delay(human_delay_speed=3)
 
@@ -498,17 +774,17 @@ class TestSelectionTreeWidget(AsyncTestCase):
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_hash_different_asset_cancel(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_cancel(
-            _get_test_data("usd/ingested_assets/output/hash_different/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/hash_different/cube.usda")
         )
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_failed_asset_cancel(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_cancel(
-            _get_test_data("usd/ingested_assets/output/ingestion_failed/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/ingestion_failed/cube.usda")
         )
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_no_metadata_asset_cancel(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_cancel(
-            _get_test_data("usd/ingested_assets/output/no_metadata/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/no_metadata/cube.usda")
         )
 
     async def __replace_mesh_ref_using_string_field_bad_ingested_asset_cancel(self, asset_path):
@@ -568,17 +844,17 @@ class TestSelectionTreeWidget(AsyncTestCase):
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_hash_different_asset_ignore(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_ignore(
-            _get_test_data("usd/ingested_assets/output/hash_different/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/hash_different/cube.usda")
         )
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_failed_asset_ignore(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_ignore(
-            _get_test_data("usd/ingested_assets/output/ingestion_failed/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/ingestion_failed/cube.usda")
         )
 
     async def test_replace_mesh_ref_using_string_field_bad_ingested_no_metadata_asset_ignore(self):
         await self.__replace_mesh_ref_using_string_field_bad_ingested_asset_ignore(
-            _get_test_data("usd/ingested_assets/output/no_metadata/cube.usda")
+            _get_test_data("usd/project_example/ingested_assets/output/no_metadata/cube.usda")
         )
 
     async def __replace_mesh_ref_using_string_field_bad_ingested_asset_ignore(self, asset_path):
@@ -633,6 +909,213 @@ class TestSelectionTreeWidget(AsyncTestCase):
 
         item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
         self.assertEquals(len(item_prims), 5)
+        await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_string_field_asset_outside_project_dir_cancel(self):
+        # Setup
+        _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+        # Select
+        usd_context = omni.usd.get_context()
+        usd_context.get_selection().set_selected_prim_paths(
+            ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+        )
+
+        await ui_test.human_delay(human_delay_speed=3)
+
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+        await item_prims[0].click()
+
+        # Make sure the mesh reference field exists
+        await ui_test.human_delay()
+        mesh_ref_field = ui_test.find(f"{_window.title}//Frame/**/StringField[*].identifier=='mesh_ref_field'")
+        self.assertIsNotNone(mesh_ref_field)
+
+        original_text = mesh_ref_field.widget.model.get_value_as_string()
+        await mesh_ref_field.click()
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.A, carb.input.KEYBOARD_MODIFIER_FLAG_CONTROL)
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.DEL)
+
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda.meta"), temp_dir)
+
+            # Attempt replacing the reference with the path of an external asset
+            await ui_test.human_delay(50)
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            await mesh_ref_field.input(asset_path, end_key=KeyboardInput.ENTER)
+            await ui_test.human_delay(human_delay_speed=3)
+
+        # Make sure that the ingestion window did not show
+        ignore_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNone(ignore_ingestion_button)
+        self.assertIsNone(cancel_ingestion_button)
+
+        # Make sure the "Copy Asset" and "Cancel" buttons exist
+        copy_external_asset_button = ui_test.find(
+            f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_external_asset_button = ui_test.find(
+            f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNotNone(copy_external_asset_button)
+        self.assertIsNotNone(cancel_external_asset_button)
+
+        # Cancel import and ensure the external asset was not imported
+        await cancel_external_asset_button.click()
+        await ui_test.human_delay(5)
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+
+        # The text should revert back to what it was originally
+        self.assertEquals(original_text, mesh_ref_field.widget.model.get_value_as_string())
+
+        await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_string_field_asset_outside_project_dir_copy_and_re_ref(self):
+        # Setup
+        _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+        # Select
+        usd_context = omni.usd.get_context()
+        usd_context.get_selection().set_selected_prim_paths(
+            ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+        )
+
+        await ui_test.human_delay(human_delay_speed=3)
+
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+        await item_prims[0].click()
+
+        # Make sure the mesh reference field exists
+        await ui_test.human_delay()
+        mesh_ref_field = ui_test.find(f"{_window.title}//Frame/**/StringField[*].identifier=='mesh_ref_field'")
+        self.assertIsNotNone(mesh_ref_field)
+
+        original_text = mesh_ref_field.widget.model.get_value_as_string()
+        await mesh_ref_field.click()
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.A, carb.input.KEYBOARD_MODIFIER_FLAG_CONTROL)
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.DEL)
+
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda.meta"), temp_dir)
+
+            # Attempt replacing the reference with the path of an external asset
+            await ui_test.human_delay(50)
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            await mesh_ref_field.input(asset_path, end_key=KeyboardInput.ENTER)
+            await ui_test.human_delay(human_delay_speed=3)
+
+            # Make sure that the ingestion window did not show
+            ignore_ingestion_button = ui_test.find(
+                f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+            )
+            cancel_ingestion_button = ui_test.find(
+                f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+            )
+            self.assertIsNone(ignore_ingestion_button)
+            self.assertIsNone(cancel_ingestion_button)
+
+            # Make sure the "Copy Asset" and "Cancel" buttons exist
+            copy_external_asset_button = ui_test.find(
+                f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+            )
+            cancel_external_asset_button = ui_test.find(
+                f"{_constants.ASSET_OUTSIDE_OF_PROJ_DIR_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+            )
+            self.assertIsNotNone(copy_external_asset_button)
+            self.assertIsNotNone(cancel_external_asset_button)
+
+            # Copy the asset
+            await copy_external_asset_button.click()
+            await ui_test.human_delay(50)
+
+            # Make sure that new ref exists
+            item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+            self.assertEqual(len(item_prims), 5)
+
+            # Check that the reference is from a new internal copy and not the original external asset
+            item_ref = item_prims[0].widget
+            self.assertEqual(item_ref.text, "cube.usda")
+            self.assertEqual(item_ref.tooltip, "./assets/ingested/cube.usda")
+
+            # Make sure the metadata matches
+            self.assertTrue(_path_utils.hash_match_metadata(file_path=asset_path, key=BASE_HASH_KEY))
+
+            # The text should not revert back to what it was originally
+            self.assertNotEquals(original_text, mesh_ref_field.widget.model.get_value_as_string())
+
+            # Delete the newly created project_example/assets/ingested subdirectory and its contents
+            shutil.rmtree(_get_test_data(f"usd/project_example/{_constants.REMIX_INGESTED_ASSETS_FOLDER}"))
+
+            await self.__destroy(_window, _selection_wid, _mesh_property_wid)
+
+    async def test_replace_mesh_ref_using_string_field_asset_outside_project_dir_no_metadata(self):
+        # Setup
+        _window, _selection_wid, _mesh_property_wid = await self.__setup_widget()  # Keep in memory during test
+
+        # Select
+        usd_context = omni.usd.get_context()
+        usd_context.get_selection().set_selected_prim_paths(
+            ["/RootNode/instances/inst_BAC90CAA733B0859_0/ref_c89e0497f4ff4dc4a7b70b79c85692da/Cube"], False
+        )
+
+        await ui_test.human_delay(human_delay_speed=3)
+
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+        await item_prims[0].click()
+
+        # Make sure the mesh reference field exists
+        await ui_test.human_delay()
+        mesh_ref_field = ui_test.find(f"{_window.title}//Frame/**/StringField[*].identifier=='mesh_ref_field'")
+        self.assertIsNotNone(mesh_ref_field)
+
+        original_text = mesh_ref_field.widget.model.get_value_as_string()
+        await mesh_ref_field.click()
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.A, carb.input.KEYBOARD_MODIFIER_FLAG_CONTROL)
+        await ui_test.emulate_keyboard_press(carb.input.KeyboardInput.DEL)
+
+        # Create a temp directory to mimic a location for an external asset
+        with tempfile.TemporaryDirectory(dir=_get_test_data("usd/")) as temp_dir:
+            shutil.copy(_get_test_data("usd/project_example/ingested_assets/output/good/cube.usda"), temp_dir)
+
+            # Attempt replacing the reference with the path of an external asset
+            await ui_test.human_delay(50)
+            asset_path = _get_test_data(f"{temp_dir}/cube.usda")
+            await mesh_ref_field.input(asset_path, end_key=KeyboardInput.ENTER)
+            await ui_test.human_delay(human_delay_speed=3)
+
+        # Make sure the ingestion window showed and that there is no ignore button
+        ignore_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='confirm_button'"
+        )
+        cancel_ingestion_button = ui_test.find(
+            f"{_constants.ASSET_NEED_INGEST_WINDOW_TITLE}//Frame/**/Button[*].name=='cancel_button'"
+        )
+        self.assertIsNone(ignore_ingestion_button)
+        self.assertIsNotNone(cancel_ingestion_button)
+
+        await cancel_ingestion_button.click()
+        await ui_test.human_delay()
+
+        # The text should revert back to what it was originally (nothing should have happened)
+        self.assertEquals(original_text, mesh_ref_field.widget.model.get_value_as_string())
+
+        # There should still be the same amount of prims
+        item_prims = ui_test.find_all(f"{_window.title}//Frame/**/Label[*].identifier=='item_prim'")
+        self.assertEquals(len(item_prims), 3)
+
         await self.__destroy(_window, _selection_wid, _mesh_property_wid)
 
     async def test_assign_single_remix_category(self):